textbox内が空の時エラーログが出る機能。 登録時に登録した日時が登録される。

+++ Gp_app/Gp_app/delete_menu.cs
@@ -1,11 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
 using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Data.SQLite;
 
@@ -23,39 +17,49 @@ namespace Gp_app
             DialogResult result = MessageBox.Show("本当に削除しますか？", "確認", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
             //ダイアログの選択結果をresultに入れる
 
-            if (result == DialogResult.Yes)//ダイアログでYesを入力したら
+            if (!String.IsNullOrEmpty(textBox1.Text))//空白じゃなかったらtrue
             {
-                using (SQLiteConnection con = new SQLiteConnection("Data Source=students.db"))
+                if (result == DialogResult.Yes)//ダイアログでYesを入力したら
                 {
-                    con.Open();
-                    using (SQLiteTransaction trans = con.BeginTransaction())
+                    using (SQLiteConnection con = new SQLiteConnection("Data Source=students.db"))
                     {
-                        SQLiteCommand cmd = con.CreateCommand();
-                        // インサート
-                        cmd.CommandText = "DELETE FROM t_product WHERE No = @no;";
-                        // パラメータセット
-                        cmd.Parameters.Add("no", DbType.Int64);
-                        // データ削除
-                        cmd.Parameters["no"].Value = int.Parse(textBox1.Text);
-                        cmd.ExecuteNonQuery();
-                        // コミット
-                        trans.Commit();
+                        con.Open();
+                        using (SQLiteTransaction trans = con.BeginTransaction())
+                        {
+                            SQLiteCommand cmd = con.CreateCommand();
+                            // インサート
+                            cmd.CommandText = "DELETE FROM t_product WHERE No = @no;";
+                            // パラメータセット
+                            cmd.Parameters.Add("no", DbType.Int64);
+                            // データ削除
+                            cmd.Parameters["no"].Value = int.Parse(textBox1.Text);
+                            cmd.ExecuteNonQuery();
+                            // コミット
+                            trans.Commit();
 
-                        this.Close();//画面を閉じる
-                        main_menu mm = new main_menu();//main_menuをmmと定義
-                        mm.Visible = true;//main_menuを表示
+                            this.Close();//画面を閉じる
+                            main_menu mm = new main_menu();//main_menuをmmと定義
+                            mm.Visible = true;//main_menuを表示
+                        }
                     }
                 }
             }
-            
+            else//空白だとfalse
+            {
+                DialogResult result2 = MessageBox.Show("数字を入力してください。", "確認", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                //ダイアログの選択結果をresultに入れる
+            }
+
+
+
         }
 
         private void button2_Click(object sender, EventArgs e)//戻るボタン
         {
-                this.Close();//画面を閉じる
-                main_menu mm = new main_menu();//main_menuをmmと定義
-                mm.Visible = true;//main_menuを表示
+            this.Close();//画面を閉じる
+            main_menu mm = new main_menu();//main_menuをmmと定義
+            mm.Visible = true;//main_menuを表示
         }
     }
-    
+
 }
